refactor(canvas): modularize node rendering and enhance spatial scaling
Decouple node rendering logic from ConceptualMapCanvas by introducing
dedicated VConceptNode and VInsightNode organisms. This migration
utilizes Vue Flow's node-types mapping to improve maintainability and
enforce Atomic Design principles.

Key technical changes:
- Relocate ConceptualMapCanvas to a pluralized directory 'canvases'.
- Implement VConceptNode and VInsightNode using VNodeContainer and
  standardized 4-port VNodeHandle configurations.
- Inherit NodeProps in custom nodes to ensure TypeScript compatibility
  with Vue Flow internal state.
- Introduce POSITION_SCALE constant to handle coordinate mapping between
  backend spatial data and frontend canvas units.
- Update ConceptualNode interface to include user_notes and status.

Breaking Change: Node rendering is now handled by type-specific
components instead of the default slot in ConceptualMapCanvas.
