+++ src/constants/canvases.ts
@@ -1,0 +1,1 @@
+export const POSITION_SCALE = 50.
